Introduce another container build stage.
This is done because nothing else I've tried to get Jenkins (Docker 1.7,
actually) not to rebuild the dependencies every time has worked. The
idea is to have a separate Jenkins job, set explicitly to run only if
things under the dependencies directory have changed, that builds a
container with the dependencies. Then the regular build will build on
top of that container.

[#132933563]